chore(session-rating): update implementation with api changes
The Lenken Server side has changed the endpoint for rating and also
the payload expected by the endpoint. This change ensures the client
makes valid calls to the server.

The Snackbar diaplay for this component was also modified to align
with the values gotten from the API.

Obeservables were also implemented to ensure that an error is actually
thrown when the request is not successful. Also, the snackbar has been
unified.