fix(date-validation): restrict date period when logging a session
Presently a user can create a session years back or in
the future by providing a past date or a future date.

This adds a validation class for the date field that
restricts the user from keying in a date beyond
the past six months or one in the future.

[Fixes #149959103]