feat(report-page): implement unmatched requests report (#223)
* feat(report-page): unmatched requests report

Currently, when a report response is sent to the frontend it does not
include a list of unmatched requests. The list of the unmatched request is
needed to facilitate displaying all unmatched requests within a given time.

This implementation refactors the report page by introducing two sub menus
placed within the report page. Because of this the routing is also updated
to allow navigation between subpages of report compoonent

Refactor the frontend implementation to include a table of unmatched
requests on the report page. The listing of requests in controlled by
pagination

[Finishes #151114935]

* chore(refactor): improve code quality
Add the `app` prefix to component selector.

Change the display of skill badges to conform to already existing
display properties, also included the date of the request to provide
more information to Admins.

Cleaned up code style issues including spaces, object key literals etc.

Reverted to the existing `body.data` usage to prevent the assumption
that all the other endpoints now don't return the `data` key.

display properties, also included the date of the request # Please enter the commit message for your changes. Lines starting