feat(all-requests-page):  allow admin close mentorship request (#226)
* feat(all-requests-page): admin can close mentorship session

Currently, it is only the mentee who can cancel a request. In cases
when a request is matched or not matched but did take place, there is
no way an admin can cancel that request.

This refactors the frontend implementation to now allow the admin to cancel
any open /matched request that was not fulfilled. A Cancel button has been added
to all open and matched request cards.

[Finishes #151641164]

* chore(refactor): clean-up style violations and restructure codebase