feat(pool-filters): add functionality for users to filter requests pool (#249)
This implementation introduces filtering functionality for the
request pool. It is crucial that users are able to filter
requests as per their preferences.

I have added a new component called pool-filters, this component
is loaded on each page where request filtering is required.

[Finishes #152160424]

* feat(pool-filter): add functionality for user to filter requests pool

This addition refactor the code to clean it.
Rename getAppliedFilters to applyFilters.
Rename getRequestsSkills to getSkillsWithRequests.

[Finished #152160424]

* style(filter-template-style): fix styles for filter template

Fix CSS styles for the filter template

[Finished #152160424]