feat(functionality): add functionality for the in-progress page (#252)
This task creates the In Progress component and its functionality.
It also adds naviagtion to the nav-bar and links to the respective pages of the `request-pool`

Using the v2 endpoint for accessing in-progress mentorship requests,
the component uses a request service method `getInProgressRequests`
and passes in query parameters userId of the current user and `status_id` of `2` in order to get those requests that are matched.

The requests records are then displayed on the `request-pool/in-progress` route.

[Delivers #152310518]

Signed-off-by: JULIE WANJA MUGIRA <julie.mugira@andela.com>

chore(styles): fix styles on the request pool page (#248)

Ensure the top navigation bar, left navigation and
right sidebar is fixed on page scroll.

[Delivers #152504159]

feat(functionality): add in progress functionality

Adds the in progress component to the request pool page.
Displays requests of the logged in user that are in-progress.

[Finishes #152310518]

Signed-off-by: JULIE WANJA MUGIRA <julie.mugira@andela.com>

feat(functionality): create the in-progress page

This task creates the in progress component and its functionality.

Using the v2 endpoint for accessing in-progress mentorship requests,
the component uses a request service method `getInProgressRequests`
and passes in query parameters userId of the current user and `status_id` of `2` in order to get those requests that are matched.

The requests records are then displayed on the `request-pool/progress` route.

[Delivers #152310518]

Signed-off-by: JULIE WANJA MUGIRA <julie.mugira@andela.com>

feat(implement): add routing to enable navigation

This commit adds routing for the request pool pages to
the nav bar component.

[Delivers #152310518]

Signed-off-by: JULIE WANJA MUGIRA <julie.mugira@andela.com>

chore(test): add tests for the in-progress page and nav-bar

This commit adds the test for the navigation bar and the in-progress
page.

This commit also changes the component name from Progress to InProgress
to reflect the in-progress page.

Signed-off-by: JULIE WANJA MUGIRA <julie.mugira@andela.com>

chore(in-progress): refactor `getInProgressRequests` method

This commit refactors the `getInProgressRequests` function
from the InProgressComponent by removing the parameters limit
and page as they are not being used anymore.

Due to the nature that the in-progress requests, they are likely to not
be many and so we don't need to set a limit or paginate the data.

Signed-off-by: JULIE WANJA MUGIRA <julie.mugira@andela.com>

chore(code-cleanup): remove unused imports

This commit removes unused imports from the `src/app/modules/request-pool/in-progress/in-progress.component.spec.ts` file

Signed-off-by: JULIE WANJA MUGIRA <julie.mugira@andela.com>

chore(code-cleanup): fix jsdoc comment

This commit fixes the jsdoc comment for the `getInprogressRequests`
requestService function to reflect that the api does not pick any
params.

Signed-off-by: JULIE WANJA MUGIRA <julie.mugira@andela.com>

chore(code-cleanup): jsdoc comments and indentation

This commit adds the jsdoc @return tag to the `getRequestsInProgress` method. It also removes unused scss and fixes identation issues.

Signed-off-by: JULIE WANJA MUGIRA <julie.mugira@andela.com>

chore(code-cleanup): fix trailing spaces and unused styling

This commit fixes trailing spaces in the `in-progress.component.html`
file. This commit also removes the margin and padding property of the body selector in the scss file because they were unused.

Signed-off-by: JULIE WANJA MUGIRA <julie.mugira@andela.com>

chore(code-cleanup): add newline at the end of file

This commit adds a newline at the end of the `router-stubs.ts` file

Signed-off-by: JULIE WANJA MUGIRA <julie.mugira@andela.com>

chore(code-cleanup): fix indentation and trailing space issues

This commit fixes indentation and trailing spaces issues
in the PR.

Signed-off-by: JULIE WANJA MUGIRA <julie.mugira@andela.com>