fix(filter-styling): fix styling for pool filters (#259)
Currently, the layout of filters elements is not properly set
as a result, the component template appears to be disorganized.

This update fixes the layout styles for pool filters so that the
layout of elements in filters template is in order. I have updated
pool-filter component scss file to apply the fix.

[Fixes #152800095]