fix(star-rating): fix single request star rating (#273)
Currently when a user clicks on a request in the request pool, a modal
pops up. The modal contains details of the request including the name
of the user who made the request and their rating. The star rating can
be amended by clicking on the stars.

This commit fixes this problem. When a user clicks on the star rating,
the value does not change.

[Fixes #153112053]