fix(notification-ui): fix notification ui design (#299)
The notification UI was not accepted due to some variance between the design and zeplin mockup.

This commit fixes the following identified issues:

The font-family of notification UI was changed to match what is on the mockup. The close X button was made a little bit larger than the notification UI title as observed in the mockup. The base of the close X button was made to be a little bit above the font baseline as observed in the mockup.

It also makes sense for the notification UI to close whenever there you click outside its borders. This feature was implemented in this commit. An invisible which listens to a click event was added to this component

The close X button and the notification icon changes the mouse pointer to hand when they are pointed to.

[Delivers #153401466]