feat(notification): add user notification to notification bar (#316)
There is already an existing UI for notification on Lenken. This commit implements notification for mentor request.

A notifyMentee method was implemented to send notifications to a mentee when there is a mentor request. In notification component, methods were implemented to, change timestamp to readable date-time, categorize notifications by date, check the type of notification an method that returns gets categorized notifications.

[Delivers #152791787]