feat(set-default-filters): set default filter to all requets (#326)
When the user navigates to the request-pool, recommended requests
are displayed and if there are no requests a blank page with
"Sadly, I have nothing to show :(." will be displayed.

This commit sets the default filter "All" requests and displays
all requests in the request-pool, if the current user has no
recommended requests.

[Delivers #154311127]