fix(select-button): fix select-button for drop-down menus (#308)
The drop-down buttons for Duration, Pairing time and Time Zone
are unresponsive in the create requests modals.

This commit creates two reusable components, skills-dropdown and
dropwdown. Skills-dropdown is used by the basic and complimentary
skills fields, and dropdown is used by the duration, pairing time and
time zone fields. The components in turn provide responsiveness when
clicking the Pairing time, Duration and Time Zone fields.

[Fixes #154029672]