fix(style): fix profile dropdown menu (#332)
This commit fixes a glitch with the profile dropdown menu on the navbar so that when a user hovers over the profile menu it doesn't bubble, this commit also fixes the notification bell image to a much clearer image.

To fix the dropdown glitch i removed previously placed divs on the markup and also the styles used to space the profile image and the arrow  so as to allow for styling with flexbox and padding positioning between the profile image and the arrow. Also i used a CSS transform property to switch between the dropdown menu arrow, this solved the problem with the bubble. I also redesign the image to give a higher pixel in the view

[Fixes #154558228]