fix(create-request-modal): fix all days check box (#318)
Currently when the all days checked box is checked and then unchecked
all the days that has been checked will not be unchecked.
And also when a user checks all days check box and uncheck one of the
days the all days check box remains checked. Lastly when a user
checks all the days the all days checked box will not be checked.

This commit ensures that when a user checks all days checkbox
and then unchecked it, all the days that has been checked should be
unchecked. And also when a user checks all days check box and
uncheck one of the days the all days checkbox will be uncheck as
well. It also ensures that whenever a user checks all the days
the all days check box will be checked as well.

[Fixes #154127673]