chore(add-scroll): make session cards scrollable (#341)
Adds the ability for the user to scroll and see files added to a
particular session, removes the scroll for the whole pending page.

This commit adds a max-height to the card class, a div containing
the card header contents is also added and the card body is
made scrollable.

[Finishes #154794357]