feat(log-session): implement the log session functionality (#333)
On lenken version 2, only logged sessions are currently displayed and
also, there is no functionality that allows mentor or mentee to log
a session. The "Log session" button responsible for this is currently
not functional.

This commit does two major things. Firstly, it displays all sessions,
both missed and the next upcoming session, and also shows to the mentee
or mentor whether they have logged or need to confirm any of the sessions.
Secondly, it adds the functionality that allows either the mentee or the
mentor to click on the "Log Session" button in any session in the
in-progress requests and the session will be logged.

[Delivers #154203217]