chore(style): align elements in the history table  (#372)
* chore(style): align elements in the history table

Currently the elements in the history table's body are misaligned with
the table's header. The ratings element has been pushed out of the table
row.

This commit fixes this by removing the table row width rule and
implementing flex's space around rule on the star rating component.

[Finishes #155209994]

* chore(pending-request-table): align columns in the pending requests table

The previous commit altered the alignment of the pending request table
due to the changes made in request-pool.scss.

This commit fixes this by reintroducing the width: 17% rule and adding
the padding
left to 2% to the pending-requests-table component-scoped
styling(pending-requests-table.scss).