fix(style): fix calendar dropdown glitch (#383)
This commit fixes a UI bug on the  within
the admin view so that when an admin user clicks on
the calendar it doesn't show behind the request pool table
header but rather on top of it.

To fix this dropdown glitch, the some components in the page
were restyled to allow for a flexed layout rather than the
previous fixed position layout thereby allowing the calendar
dropdown to stay on top of the page. The request pool header
style was refactored to use a absolute positioning rather than
a fixed positioning and to get the header fixed the container div
was flexed to grow with the remaining space of the page thereby making
the header fixed, also a spacer was used to get the first result visible
to the user. Other pages such as skills and report whose styling
where affect by these changes where also addressed.

[Fixes #155323508]