fix(exporting-csv): fix dates on csv filenames and disable export button (#384)
Currently the filenames being generated for the csv files only show
January as the month for their startDates and endDates. There is also no
indication to a user if the export process is ongoing.

This commit rectifies this by removing a typo in the formatDate method
on the requestStatistics component. It also creates a component
called ExportButtonComponent to control the disabling of the export button.

[Delivers #155395998]