feat(create-request-modal): make view button functional (#394)
Currently, after a user has created a request successfully, a modal
is displayed with two buttons labelled 'close' and 'view'. The 'view' button
is not functional.

This commit ensures that whenever a user clicks on the view button, a modal is
popped up to the user that displays the details of the request that has just been
created by the user.

* chore(refactor): transfer responsibility for create confirmation modal

The scope of responsibility for displaying as request details after
creation when a user clicks "View" has been moved from the header to
the create request component. This will help provide context to the
confirmation modal.

Also, the ConfirmationAlertComponent didn't implement afterClose for
when a user closes the modal without using the confirmAction. This has
been rectified to it also calls afterClose for the close action directly.

The TSLint config was changed to specify 2 indents also.

[Delivers #155463527]