feat(confirm-session): implement confirm session functionality (#393)
Currently, users cannot confirm logged sessions because
the confirm session button does nothing.

This commit implements the functionality that allows users
to confirm logged sessions. When the user clicks on the confirm
session button, a modal (with information about the session)
opens.

When the user confirms the session, the mentor_approved or the
mentee_approved field in the database is updated. Ratings
and comment are also sent to the backend if provided.

[Delivers #154205554]