chore(dashboard-requests): ensure the dashboard shows all requests (#392)
* chore(dashboard-requests): ensure the dashboard still shows all requests

The Admin Dashboard `Requests` Page currently gets the
`mentorship requests` it displays from the `GET: /api/v2/requests/pool`
endpoint, this endpoint is being changed, as such, it would not return
all mentorship requests again.

This would break the `Admin Dashboard` from displaying all requests,
it is therefore important that the endpoint consumed by the
Admin Dashboard Requests page is changed to one that returns
`All Mentorship Requests`

This commit, therefore, ensures that the Admin Dashboard `Requests
Page` is made to get all the mentorship requests by connecting
it to consume another endpoint namely
`GET: /api/v2/requests/all-requests`

* chore(refactor): remove code duplication

The PoolComponent can make one of two API calls to get requests. This
is determined by parent component including a boolean flag to determine
if all requests should be gotten or if the requests a use has interests
in should be excluded.

The data gotten from both API calls has the same structure and the
actions on the data is the same. Rather than duplicate the code that
resolves each promise, I wrapped both within a promise and resolve it
with one of the two responses. This allows me use just on call to act
on the data gotten from the API.

Also, I renamed the getAllRequests to be the root getRequests as it's
the standard for REST APIs and the one specific to pool requests was
renamed to fit the responsibility.

[Finishes #155607109]