chore(lazy-load-modules): lazy-load all-requests and reports modules (#423)
Currently, all-requests and reports modules are loaded
when the application starts. This has a negative effect
on the speed of the application. It is ideal that these
modules are not loaded until there is a need for them
in the application.

This commit ensures that the two modules stated above
 are only loaded when the application needs them by
removing the two modules from app.module.ts so that
they are lazy-loaded when needed.

[Finishes #155964646]