feat(add-file-modal): implement file upload with file browser window (#425)
Currently files can only be uploaded via a drag and drop action on the
file upload box. This commit allows users to click the upload box to
open a web file browser window to choose the file they want to
upload.

A file input html element has been added to hanlde this. The visibility
of this element has been hidden. When a user chooses his file the
onchange event on the file input adds the file to the component for it
to be uploaded.

[Finishes #155844539]