chore(create-request): remove create request add skill button (#449)
Currently the create request form contains a add button to add skills the user wishes to learn/tutor. This button is usually missed by users hence the user is unable to submit the form as the field is required for the form to be submitted.

This commit replaces this button with an auto adding form field. When a skill is clicked from the drop down list, the user no longer needs to click the add button to add the skill to the list of skill they want to pick as this is added on selected.

[Finishes #156535683]