feat(ratings): implement mentor/mentee rating (#447)
Currently, on the lenken app, only a mentee can rate
a mentor. Also, user is not able to edit a logged
session.

This commit implement the mentor/mentee rating by
refactoring the LogSession class to accomodate a
mentor rate a mentee. This commit also implement
editing a session that is yet to be confirmed by
either mentor/mentee.

This commit also implement rejecting an already
logged session.

[Delivers #156301497]