fix(endless-scroll): fix endless scroll (#460)
The endless scroll doesn't function as expected. When a user scrolls
down to the bottom of the page, no new requests are displayed.

This commit fixes the endless scroll by removing the imported infinite
scroll module and adds an infinite scroll directive. The directive is
triggered in the pool page div when a user scrolls down to 70% of the
current requests.This makes an API call and fetches more requests
from the API.

[Fixes #154760521]